refactor: unified warehouse tests
